// week-4/reactive-composer-app/src/app/composer.interface.ts
/*
================================================================
    Title: composer.interface.ts
    Author: Carl Logan
    Date: 1/21/2023
    Description: Passing Data to Routes, Part 1.
================================================================
*/

export interface IComposer {
  composerId: number;
  fullName: string;
  genre: string;
}